fix(internal): moving course middlewares and server to shared folder
