+++ cmd/server/main.go
@@ -10,9 +10,9 @@ import (
 	"github.com/gorilla/mux"
 	_ "github.com/lib/pq"
 	"github.com/sumelms/microservice-course/internal/course"
-	"github.com/sumelms/microservice-course/internal/course/transport/http"
 	"github.com/sumelms/microservice-course/internal/matrix"
 	"github.com/sumelms/microservice-course/internal/matrix/clients"
+	http "github.com/sumelms/microservice-course/internal/shared"
 	"github.com/sumelms/microservice-course/pkg/config"
 	database "github.com/sumelms/microservice-course/pkg/database/postgres"
 	log "github.com/sumelms/microservice-course/pkg/logger"
