+++ internal/shared/server.go
